requirements.txt
fastapi
uvicorn[standard]
pytest
pytest-cov
httpx
python-dateutil
pydantic-settings
